requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
fastapi==0.115.5
fire==0.7.0
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
matplotlib==3.9.2
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==11.0.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.13
pyparsing==3.2.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.1.0
tqdm==4.67.0
urllib3==2.2.3
uvicorn==0.32.0
websockets==14.1
